Add missing components i.e., BookList and CreateNewBook
