+++ src/components/books/BookList.js
@@ -1,0 +1,12 @@
+const BookList = () => {
+  return (
+    <ul>
+      <li>Seven Habits of Highly Effective People ~ Unknown </li>
+      <button type="button">Remove</button>
+      <li>Rich Dad, Poor Dad ~ Unknown</li>
+      <button type="button">Remove</button>
+    </ul>
+  );
+};
+
+export default BookList;
