+++ src/components/books/CreateNewBook.js
@@ -1,0 +1,11 @@
+const CreateNewBook = () => {
+  return (
+    <form>
+      <input type="text" />
+      <input type="text" />
+      <input type="submit" value="Submit" />
+    </form>
+  );
+};
+
+export default CreateNewBook;
