+++ src/components/books/books.js
@@ -1,0 +1,11 @@
+import BookList from './BookList';
+import CreateNewBook from './CreateNewBook';
+
+const Books = () => (
+  <section>
+    <BookList />
+    <CreateNewBook />
+  </section>
+);
+
+export default Books;
