+++ src/components/categories/categories.js
@@ -1,0 +1,3 @@
+const Categories = () => <h1>Under Construction</h1>;
+
+export default Categories;
