--- src/redux/books/books.js
@@ -1,17 +1,0 @@
-const Books = () => (
-  <section>
-    <ul>
-      <li>Seven Habits of Highly Effective People ~ Unknown </li>
-      <button type="button">Remove</button>
-      <li>Rich Dad, Poor Dad ~ Unknown</li>
-      <button type="button">Remove</button>
-    </ul>
-    <form>
-      <input type="text" />
-      <input type="text" />
-      <input type="submit" value="Submit" />
-    </form>
-  </section>
-);
-
-export default Books;
