--- src/redux/categories/categories.js
@@ -1,3 +1,0 @@
-const Categories = () => <h1>Under Construction</h1>;
-
-export default Categories;
